refactor: AI 프롬프트 관리 로직 분리 및 공통 지침(Global Guide) 주입
